+++ BossMod/Modules/RealmReborn/Alliance/A12Thanatos/A12Thanatos.cs
@@ -11,8 +11,8 @@ public class A12Thanatos(WorldState ws, Actor primary) : BossModule(ws, primary,
     protected override void DrawEnemies(int pcSlot, Actor pc)
     {
         Arena.Actor(PrimaryActor);
-        Arena.Actors(Enemies(OID.MagicPot), Colors.Object);
-        Arena.Actors(Enemies(OID.Nemesis));
-        Arena.Actors(Enemies(OID.Sandman));
+        Arena.Actors(Enemies((uint)OID.MagicPot), Colors.Object);
+        Arena.Actors(Enemies((uint)OID.Nemesis));
+        Arena.Actors(Enemies((uint)OID.Sandman));
     }
 }
